mali: Stop using the reserved-memory region to setup the fb area
The mali driver implements a security check by having the platform code set
up the framebuffer start address and its size. This was implemented as a
reserved memory region to not give access to all the memory region.

However, this has a bunch of drawbacks:
  - Every user has to create a reserved memory region in the DT, that
    cannot be upstreamed.
  - Even with that check in place, if we don't want to just waste RAM, we
    could have some movable pages (such as caches) that could leak, so
    that's not a great protection either.
  - It prevents the build of the module without the kernel source.
  - It creates another CMA region that takes RAM away.
  - It generates a lot of questions from users.

Let's remove this, and the matching documentation.

Signed-off-by: Maxime Ripard <maxime.ripard@bootlin.com>